orm testing branch

+++ src/main/java/com/room6/student_tutor/data/ForumRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ForumRepository extends CrudRepository<Forum, Integer> {
+public interface ForumRepository extends CrudRepository<Forum, Long> {
 }
